+++ src/opencl/implementations/uint256/subtraction.cpp
@@ -5,7 +5,7 @@ void uint256Subtraction(const UInt256 *a, const UInt256 *b, UInt256 *result)
 {
 
   result->limbs[3] = a->limbs[3] - b->limbs[3];
-  unsigned long borrow = a->limbs[3] < b->limbs[3];
+  unsigned int borrow = a->limbs[3] < b->limbs[3];
 
   result->limbs[2] = a->limbs[2] - b->limbs[2] - borrow;
   borrow = (a->limbs[2] < b->limbs[2]) | ((a->limbs[2] == b->limbs[2]) & borrow);
